+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+- Added ability to attach controllers of different types to the same widget
+- Added ability to manually detach a controller from a widget
+- Added documentation comments
+- Updated README.md with a short example
+- Added example.dart
+
 ## 1.0.0
 
 - Added `TurboController`